Restore relevant function when exporting html
# What
Remove previously modified custom script imports, restore the original
inline style and restore css-related import function.

# Why
Currently the javascript snippet injected in the html is interfering
with the MathJax context menu, and the user wants to add additional css
styles to the exported html file, so the functionality is restored.